fix: split builder.rs into builder.rs + builder_core.rs to satisfy ≤500 LOC invariant
